Refactor PrintHelpCommand:
Parser should not know what exact help message to print. Instead it should just parse the user input and call the corresponding commands. Hence, the specific messages to be printed are now handled by the PrintHelpCommand.
